% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_emdexp}
\alias{calculate_emdexp}
\title{Earth Mover's Distance for differential expression analysis}
\usage{
calculate_emdexp(expM, samplesA, samplesB, binSize = 0.2, nperm = 100,
  verbose = TRUE)
}
\arguments{
\item{expM}{A gene expression matrix. The rownames should contain gene
identifiers, while the column names should contain sample identifiers.}

\item{samplesA}{A vector of sample names identifying samples in \code{expM}
that belong to "group A". The names must corresponding to column names
in \code{expM}.}

\item{samplesB}{A vector of sample names identifying samples in \code{expM}
that belong to "group B". The names must corresponding to column names
in \code{expM}.}

\item{binSize}{The bin size to be used when generating histograms of
gene expression levels for "group A" and "group B".}

\item{nperm}{An integer specifying the number of randomly permuted EMD
scores to be computed. Defaults to 100.}

\item{verbose}{Boolean specifying whether to display progress messages.}
}
\value{
The function returns an \code{\link{emdexp}} object.
}
\description{
This is the main user interface to the \pkg{emdexp} package, and is
usually the only function needed.
}
\details{
details go here
}
\examples{
foo <- 1
}
\seealso{
\code{\link{emdexp}} \code{\link[emdist]{emd2d}}
}

