% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateEMDExp}
\alias{calculateEMDExp}
\title{Earth Mover's Distance for differential expression analysis}
\usage{
calculateEMDExp(expM1, expM2, expM = NA, samplesA = NA, samplesB = NA,
  binSize = 0.2, nperm = 100, stepSize = 0.001, verbose = TRUE)
}
\arguments{
\item{expressionData}{foo}
}
\value{
The function returns a list with the following elements:
something...
}
\description{
This is the main user interface to the \pkg{emdexp} package, and is
usually the only function needed.
}
\details{
details go here
}
\examples{
foo <- 1
}
\seealso{
\code{\link[emdist]{emd2d}}
}

