% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateEMDfoo}
\alias{calculateEMDfoo}
\title{Compute Earth Mover's Distance}
\usage{
calculateEMDfoo(expressionData, samplesA, samplesB, binSize = 0.2,
  nperm = 1000, verbose = TRUE)
}
\arguments{
\item{expressionData}{foo}
}
\value{
The function returns a list with the following elements:
something...
}
\description{
This is the main user interface to the \pkg{emd} package, and is
usually the only function needed.
}
\details{
details go here
}
\examples{
foo <- 1
}
\seealso{
\code{\link[emdist]{emd2d}}
}

