% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateGeneEMD}
\alias{calculateGeneEMD}
\title{Compute Earth Mover's Distance for a single gene}
\usage{
calculateGeneEMD(geneData, idxA, idxB, binSize = 0.2)
}
\arguments{
\item{expressionData}{foo}
}
\value{
return...
}
\description{
description..
}
\details{
details...
}
\examples{
foo <- 1
}
\seealso{
\code{\link{calculateEMD}} \code{\link[emdist]{emd2d}}
}

