\name{EMD-internal}
\alias{EMD-internal}
\alias{hilbert}
\alias{unwrap}
\alias{C_extrema2dC}
\alias{extrema2dVC}
\alias{imageplot.ts}
\title{Internal EMD Functions}
\description{
  The \code{hilbert}, \code{unwrap}, \code{C_extrema2dC}, \code{extrema2dVC} and \code{imageplot.ts} are internal EMD functions.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
