% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{emdexp-package}
\alias{emdexp-package}
\title{Implements the Earth Mover's Distance algorithm for analyzing
differential expression of heterogenous data sets.}
\description{
\code{\link{calculate_emdexp}} will usually be the only function needed.
}
\references{
ref to paper goes here
}

