% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_density}
\alias{plot_density}
\title{Plot expression distributions and emd score for a gene.}
\usage{
plot_density(emdexp, gene_name)
}
\arguments{
\item{emdexp}{An \code{\link{emdexp}} object, typically returned via a call
to \code{\link{calculate_emdexp}}.}

\item{gene_name}{The gene to visualize. The name should be defined as a row
name in \code{emdexp$emd}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object is returned. If the value is
not assigned, a plot will be drawn.
}
\description{
The expression data for the specified gene is retrieved from
\code{emdexp$emd}. \code{emdexp$samplesA} and \code{emdexp$samplesB} are used
to divide the data into two distributions, which are then visualized as
density distributions. The calculated emd score for the specified gene is
displayed in the plot title.
}
\seealso{
\code{\link{calculate_emdexp}} \code{\link[ggplot2]{ggplot}}
}

