% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_emdnull}
\alias{plot_emdnull}
\title{Plot null distribution of permuted emd scores vs. calculated emd
scores.}
\usage{
plot_emdnull(emdexp)
}
\arguments{
\item{emdexp}{An \code{\link{emdexp}} object, typically returned via a call
to \code{\link{calculate_emdexp}}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object is returned. If the value is
not assigned, a plot will be drawn.
}
\description{
The median of the randomly permuted emd scores (i.e. the null
distribution) is plotted on the x-axis, vs. the observed emd scores on the
y-axis. The line \code{y=x} is superimposed.
}
\seealso{
\code{\link{calculate_emdexp}} \code{\link[ggplot2]{ggplot}}
}

