% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_perms}
\alias{plot_perms}
\title{Plot histogram of emd values calculated via random permutation.}
\usage{
plot_perms(emdexp)
}
\arguments{
\item{emdexp}{An \code{\link{emdexp}} object, typically returned via a call
to \code{\link{calculate_emdexp}}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object is returned. If the value is
not assigned, a plot will be drawn.
}
\description{
The permuted emd scores stored in \code{emdexp$emd.perm} are
plotted as a histogram.
}
\seealso{
\code{\link{calculate_emdexp}} \code{\link[ggplot2]{ggplot}}
}

