% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EmdExp}
\alias{EmdExp}
\title{Create an EMD object}
\usage{
EmdExp(expM, samplesA, samplesB, emd, emd.perm)
}
\arguments{
\item{expressionData}{foo}
}
\value{
The function returns a list with the following elements:
something...
}
\description{
description here
}
\details{
details go here
}

